trying refactoring
